* Use bytes.Buffer instead of io.Writer in some functions
* Use less fmt
* Add benchmarks

```
benchmark                        old ns/op     new ns/op     delta
BenchmarkGenManToFile-4          30037         29001         -3.45%
BenchmarkGenMarkdownToFile-4     23096         10295         -55.43%
BenchmarkGenYamlToFile-4         62488         61622         -1.39%

benchmark                        old allocs     new allocs     delta
BenchmarkGenManToFile-4          77             89             +15.58%
BenchmarkGenMarkdownToFile-4     47             48             +2.13%
BenchmarkGenYamlToFile-4         121            120            -0.83%

benchmark                        old bytes     new bytes     delta
BenchmarkGenManToFile-4          9075          9651          +6.35%
BenchmarkGenMarkdownToFile-4     1984          2832          +42.74%
BenchmarkGenYamlToFile-4         26706         26280         -1.60%
```